actualización 23 Dic: * Se crea proyectos.js, para manejar solo la lógica de proyectos * Se crea tablaGenerica.js para manejar eventos de un control de table
