+++ principal.php
@@ -50,6 +50,7 @@ else {
         <script src="https://cdn.datatables.net/select/1.2.2/js/dataTables.select.min.js"></script>
         <script src="https://code.jquery.com/ui/1.12.0/jquery-ui.min.js" ></script>
 
+        
 <!-- 
         <script src="https://code.jquery.com/ui/1.13.2/jquery-ui.js"></script>
         <link rel="stylesheet" href="//code.jquery.com/ui/1.13.2/themes/base/jquery-ui.css">
